+++ rtos/rtos.h
@@ -18,7 +18,7 @@
 void OS_Init(uint32_t slice, triggerType source);
 void OS_Enable(void);
 void OS_Disable(void);
-int OS_Task_Create(voidfuncptr task_entry, void *argv);
-int OS_Task_Delete(voidfuncptr task);
+uint8_t OS_Task_Create(voidfuncptr task_entry, void *argv);
+uint8_t OS_Task_Delete(voidfuncptr task);
 
 #endif /* RTOS_H_ */
